# truck-service-host/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Truck Service API Documentation",
        "description": "Interactive API documentation for the Truck Service application. Developed by Mohammed Sleem.",
        "contact": {
            "email": "dev.mohammed@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://bull-station.com",
            "description": "Main API Server"
        }
    ],
    "paths": {
...
        "/api/my-bookings": {
            "get": {
                "tags": [
                    "Booking Management"
                ],
                "summary": "Get the authenticated user's bookings",
                "operationId": "getMyBookings",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "approved",
                                "confirmed",
                                "rejected",
                                "cancelled",
                                "completed"
                            ]
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "incoming",
                                "outgoing"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of bookings for the authenticated user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/BookingResource"
                                            }
                                        },
                                        "links": {
                                            "type": "object",
                                            "example": {
                                                "first": "url",
                                                "last": "url",
                                                "prev": null,
                                                "next": "url"
                                            }
                                        },
                                        "meta": {
                                            "type": "object",
                                            "example": {
                                                "current_page": 1,
                                                "from": 1,
                                                "last_page": 10,
                                                "path": "url",
                                                "per_page": 15,
                                                "to": 15,
                                                "total": 150
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "status": [
                                                    "The selected status is invalid."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/bookings": {
            "post": {
                "tags": [
                    "Booking Management"
                ],
                "summary": "Create a new booking request",
                "operationId": "storeBooking",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "truck_id",
                                    "start_datetime",
                                    "end_datetime",
                                    "days",
                                    "hours",
                                    "needs_delivery"
                                ],
                                "properties": {
                                    "truck_id": {
                                        "type": "integer"
                                    },
                                    "start_datetime": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-10-20 08:00:00"
                                    },
                                    "end_datetime": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-10-22 17:00:00"
                                    },
                                    "days": {
                                        "type": "integer",
                                        "example": "2"
                                    },
                                    "hours": {
                                        "type": "integer",
                                        "example": "9"
                                    },
                                    "needs_delivery": {
                                        "type": "boolean",
                                        "example": "true"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Booking request sent successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Booking request sent successfully. Awaiting owner approval."
                                        },
                                        "booking_id": {
                                            "type": "integer",
                                            "example": 123
                                        },
                                        "booking_details": {
                                            "properties": {
                                                "total_price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 350.5
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "pending"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict, truck not available",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, this truck is no longer available for the selected dates."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "truck_id": [
                                                    "The selected truck id is invalid."
                                                ],
                                                "start_datetime": [
                                                    "The start datetime is not a valid date."
                                                ],
                                                "end_datetime": [
                                                    "The end datetime must be a date after or equal to start datetime."
                                                ],
                                                "days": [
                                                    "The days must be at least 0."
                                                ],
                                                "hours": [
                                                    "The hours must be at least 0."
                                                ],
                                                "needs_delivery": [
                                                    "The needs delivery field must be true or false."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/my-bookings/{booking}/approve": {
            "post": {
                "tags": [
                    "Booking Management"
                ],
                "summary": "Approve a booking request (Truck Owner)",
                "operationId": "approveBooking",
                "parameters": [
                    {
                        "name": "booking",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Booking has been approved. Awaiting payment.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Booking has been approved. Awaiting payment."
                                        },
                                        "booking": {
                                            "$ref": "#/components/schemas/BookingResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "This action is unauthorized."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/my-bookings/{booking}/reject": {
            "post": {
                "tags": [
                    "Booking Management"
                ],
                "summary": "Reject a booking request (Truck Owner)",
                "operationId": "rejectBooking",
                "parameters": [
                    {
                        "name": "booking",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Booking has been rejected.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Booking has been rejected."
                                        },
                                        "booking": {
                                            "$ref": "#/components/schemas/BookingResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "This action is unauthorized."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "MyTruckResource": {
                "title": "My Truck Resource",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "model": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "active",
                            "inactive"
                        ]
                    },
                    "price_per_day": {
                        "type": "string",
                        "format": "decimal"
                    },
                    "category": {
                        "type": "string"
                    },
                    "main_image": {
                        "type": "string",
                        "format": "url"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "BookingResource": {
                "title": "Booking Resource",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "status": {
                        "type": "string"
                    },
                    "start_datetime": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "end_datetime": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "total_price": {
                        "type": "string",
                        "format": "decimal"
                    },
                    "truck": {
                        "$ref": "#/components/schemas/MyTruckResource"
                    },
                    "other_party": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}